+++ src/domains/auth/auth.controller.ts
@@ -1,27 +1,13 @@
-import bcrypt from 'bcrypt';
 import { Request, Response } from 'express';
 
-import { User } from './auth';
-
-let userId = 1;
-const userRepository: User[] = [];
+import { UserData } from './auth';
+import * as authService from './auth.service';
 
 
 export const doLogin = async (req: Request, res: Response) => {
-    const email = req.body.email;
-    const password = req.body.password;
+    const userData: UserData = req.body;
 
-    const user: User | undefined = userRepository.find(user => user.email = email);
-
-    if(!user) {
-        return new Error('Not Found User');
-    }
-
-    const isMatch = await bcrypt.compare(password, user.password);
-
-    if( !isMatch ) {
-        return new Error('Invalid User Information') // 유저 비밀번호 오류 처리 필요
-    }
+    const user = await authService.doLogin(userData);
 
     // Type Error 수정 필요
     req.session.isLoggedIn = true;
@@ -42,15 +28,10 @@ export const doLogin = async (req: Request, res: Response) => {
 export const doSignup = async (req: Request, res: Response) => {
 
     // type safe를 위해 추후 validation 추가 예정
-    const email: string = req.body.email;
-    const password: string = req.body.password;
+    const userData: UserData = req.body;
 
-    const hashPassword = await bcrypt.hash(password, 12);
+    const user = await authService.doSignup(userData);
 
-    const user: User = {id: userId, email: email, password: hashPassword };
-    userRepository.push(user);
-    userId++;
-
-    res.status(201).json({id: user.id, email: email});
+    res.status(201).json({id: user.id, email: user.email});
 
 }
